requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
pydantic==2.6.1
python-multipart==0.0.6
python-dotenv==1.0.1
requests==2.31.0
aiofiles==23.2.1
httpx==0.26.0

# VideoDB dependencies
videodb==0.2.3

# OpenAI - only AI provider
openai==1.3.7

# Additional utilities
tqdm==4.66.1
Pillow==10.1.0
annotated-types==0.6.0
anyio>=3.7.1,<4.0.0
backoff==2.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.0
filelock==3.13.1
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.2
idna==3.6
packaging==23.2
pydantic_core==2.16.2
PyYAML==6.0.1
sniffio==1.3.0
tokenizers==0.15.1
typing_extensions==4.9.0
urllib3==2.2.0
